[release-v1.1] Get external topics (#143)
* :lipstick: prefactoring for topic reconciling and finalizing (#1982)


* :magic_wand: Adding private annotation to drive the name of the backing topic

+++ control-plane/pkg/reconciler/broker/broker.go
@@ -48,6 +48,10 @@ import (
 const (
 	// TopicPrefix is the Kafka Broker topic prefix - (topic name: knative-broker-<broker-namespace>-<broker-name>).
 	TopicPrefix = "knative-broker-"
+
+	// private annotation for changing the topic
+	// NOTE: this may go away in a future release
+	externalTopicAnnotation = "x-kafka.eventing.knative.dev/external.topic"
 )
 
 type Reconciler struct {
@@ -122,26 +126,10 @@ func (r *Reconciler) reconcileKind(ctx context.Context, broker *eventing.Broker)
 		return fmt.Errorf("failed to track secret: %w", err)
 	}
 
-	topicName := kafka.BrokerTopic(TopicPrefix, broker)
-
-	saramaConfig, err := kafka.GetSaramaConfig(securityOption)
+	topic, err := r.reconcileBrokerTopic(broker, securityOption, statusConditionManager, topicConfig, logger)
 	if err != nil {
-		return statusConditionManager.FailedToCreateTopic(topicName, fmt.Errorf("error getting cluster admin config: %w", err))
+		return err
 	}
-
-	kafkaClusterAdminClient, err := r.NewKafkaClusterAdminClient(topicConfig.BootstrapServers, saramaConfig)
-	if err != nil {
-		return statusConditionManager.FailedToCreateTopic(topicName, fmt.Errorf("cannot obtain Kafka cluster admin, %w", err))
-	}
-	defer kafkaClusterAdminClient.Close()
-
-	topic, err := kafka.CreateTopicIfDoesntExist(kafkaClusterAdminClient, logger, topicName, topicConfig)
-	if err != nil {
-		return statusConditionManager.FailedToCreateTopic(topic, err)
-	}
-	statusConditionManager.TopicReady(topic)
-
-	logger.Debug("Topic created", zap.Any("topic", topic))
 
 	// Get contract config map.
 	contractConfigMap, err := r.GetOrCreateDataPlaneConfigMap(ctx)
@@ -241,6 +229,46 @@ func (r *Reconciler) reconcileKind(ctx context.Context, broker *eventing.Broker)
 	return nil
 }
 
+func (r *Reconciler) reconcileBrokerTopic(broker *eventing.Broker, securityOption kafka.ConfigOption, statusConditionManager base.StatusConditionManager, topicConfig *kafka.TopicConfig, logger *zap.Logger) (string, reconciler.Event) {
+
+	saramaConfig, err := kafka.GetSaramaConfig(securityOption)
+	if err != nil {
+		return "", statusConditionManager.FailedToResolveConfig(fmt.Errorf("error getting cluster admin config: %w", err))
+	}
+
+	kafkaClusterAdminClient, err := r.NewKafkaClusterAdminClient(topicConfig.BootstrapServers, saramaConfig)
+	if err != nil {
+		return "", statusConditionManager.FailedToResolveConfig(fmt.Errorf("cannot obtain Kafka cluster admin, %w", err))
+	}
+	defer kafkaClusterAdminClient.Close()
+
+	// if we have a custom topic annotation
+	// the topic is externally manged and we do NOT need to create it
+	topicName, externalTopic := isExternalTopic(broker)
+	if externalTopic {
+		isPresentAndValid, err := kafka.AreTopicsPresentAndValid(kafkaClusterAdminClient, topicName)
+		if err != nil {
+			return "", statusConditionManager.TopicsNotPresentOrInvalidErr([]string{topicName}, err)
+		}
+		if !isPresentAndValid {
+			// The topic might be invalid.
+			return "", statusConditionManager.TopicsNotPresentOrInvalid([]string{topicName})
+		}
+	} else {
+		// no external topic, we create it
+		topicName = kafka.BrokerTopic(TopicPrefix, broker)
+
+		topic, err := kafka.CreateTopicIfDoesntExist(kafkaClusterAdminClient, logger, topicName, topicConfig)
+		if err != nil {
+			return "", statusConditionManager.FailedToCreateTopic(topic, err)
+		}
+	}
+
+	statusConditionManager.TopicReady(topicName)
+	logger.Debug("Topic created", zap.Any("topic", topicName))
+	return topicName, nil
+}
+
 func (r *Reconciler) FinalizeKind(ctx context.Context, broker *eventing.Broker) reconciler.Event {
 	return retry.RetryOnConflict(retry.DefaultBackoff, func() error {
 		return r.finalizeKind(ctx, broker)
@@ -325,34 +353,47 @@ func (r *Reconciler) finalizeKind(ctx context.Context, broker *eventing.Broker) 
 
 	// get security option for Sarama with secret info in it
 	securityOption := security.NewSaramaSecurityOptionFromSecret(secret)
-
-	saramaConfig, err := kafka.GetSaramaConfig(securityOption)
-	if err != nil {
-		// even in error case, we return `normal`, since we are fine with leaving the
-		// topic undeleted e.g. when we lose connection
-		return fmt.Errorf("error getting cluster admin sarama config: %w", err)
-	}
-
-	kafkaClusterAdminClient, err := r.NewKafkaClusterAdminClient(topicConfig.BootstrapServers, saramaConfig)
-	if err != nil {
-		// even in error case, we return `normal`, since we are fine with leaving the
-		// topic undeleted e.g. when we lose connection
-		return fmt.Errorf("cannot obtain Kafka cluster admin, %w", err)
-	}
-	defer kafkaClusterAdminClient.Close()
-
-	topic, err := kafka.DeleteTopic(kafkaClusterAdminClient, kafka.BrokerTopic(TopicPrefix, broker))
-	if err != nil {
+	if err := r.finalizeBrokerTopic(broker, securityOption, topicConfig, logger); err != nil {
 		return err
 	}
-
-	logger.Debug("Topic deleted", zap.String("topic", topic))
 
 	// TODO(pierDipi) remove after some releases (released in 1.4)
 	if err := r.removeFinalizerCM(ctx, finalizerCM(broker), brokerConfig); err != nil {
 		return err
 	}
 
+	return nil
+}
+
+func (r *Reconciler) finalizeBrokerTopic(broker *eventing.Broker, securityOption kafka.ConfigOption, topicConfig *kafka.TopicConfig, logger *zap.Logger) reconciler.Event {
+
+	// External topics are not managed by the broker,
+	// therefore we do not delete them
+	_, externalTopic := isExternalTopic(broker)
+	if !externalTopic {
+		saramaConfig, err := kafka.GetSaramaConfig(securityOption)
+		if err != nil {
+			// even in error case, we return `normal`, since we are fine with leaving the
+			// topic undeleted e.g. when we lose connection
+			return fmt.Errorf("error getting cluster admin sarama config: %w", err)
+		}
+
+		kafkaClusterAdminClient, err := r.NewKafkaClusterAdminClient(topicConfig.BootstrapServers, saramaConfig)
+		if err != nil {
+			// even in error case, we return `normal`, since we are fine with leaving the
+			// topic undeleted e.g. when we lose connection
+			return fmt.Errorf("cannot obtain Kafka cluster admin, %w", err)
+		}
+		defer kafkaClusterAdminClient.Close()
+
+		topic, err := kafka.DeleteTopic(kafkaClusterAdminClient, kafka.BrokerTopic(TopicPrefix, broker))
+		if err != nil {
+			return err
+		}
+
+		logger.Debug("Topic deleted", zap.String("topic", topic))
+		return nil
+	}
 	return nil
 }
 
@@ -482,4 +523,9 @@ func (r *Reconciler) removeFinalizerCM(ctx context.Context, finalizer string, cm
 
 func finalizerCM(object metav1.Object) string {
 	return fmt.Sprintf("%s/%s-%s", "kafka.brokers.eventing.knative.dev", object.GetNamespace(), object.GetName())
+}
+
+func isExternalTopic(broker *eventing.Broker) (string, bool) {
+	topicAnnotationValue, ok := broker.Annotations[externalTopicAnnotation]
+	return topicAnnotationValue, ok
 }
